requirements.txt
flask==3.0.0
flask-login==0.6.3
pymongo==4.6.0
werkzeug==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
youtube-search-python==1.6.6
wikipedia==1.4.0
openai==1.6.1
python-dotenv==1.0.0
nltk==3.8.1
google-generativeai==0.3.2
sentence-transformers==2.2.2

